moved DamageContext from impl to api
